move classifyAppStack
