+++ packages/benchmarks/src/app-development/AppDevelopmentEval.ts
@@ -6,7 +6,7 @@ import {
 } from "mongodb-rag-core/braintrust";
 import { ConversationEvalCase } from "mongodb-rag-core/eval";
 import { LlmOptions } from "mongodb-rag-core/executeCode";
-import { AppStackClassification } from "./metrics/classifyAppStack";
+import { AppStackClassification } from "./classifyAppStack";
 
 export type AppDevelopmentEvalCaseInput = {
   name: string;
